Split: take away module from __init__.py
