+++ py3status/__init__.py
@@ -2,13 +2,11 @@ from __future__ import print_function
 
 import argparse
 import ast
-import imp
 import locale
 import os
 import pkgutil
 import sys
 
-from collections import OrderedDict
 from copy import deepcopy
 from datetime import datetime, timedelta
 from json import dumps, loads
@@ -27,6 +25,7 @@ from syslog import syslog, LOG_ERR, LOG_INFO, LOG_WARNING
 from py3status.profiling import profile
 from py3status.helpers import jsonify, print_line, print_stderr
 from py3status.events import IOPoller, Events
+from py3status.module import Module
 
 try:
     from setproctitle import setproctitle
@@ -557,231 +556,6 @@ class I3status(Thread):
         self.last_output_ts = datetime.utcnow()
         self.last_prefix = ','
         self.update_json_list()
-
-
-class Module(Thread):
-    """
-    This class represents a user module (imported file).
-    It is reponsible for executing it every given interval and
-    caching its output based on user will.
-    """
-
-    def __init__(self, lock, config, module, i3_thread, user_modules):
-        """
-        We need quite some stuff to occupy ourselves don't we ?
-        """
-        Thread.__init__(self)
-        self.click_events = False
-        self.config = config
-        self.has_kill = False
-        self.i3status_thread = i3_thread
-        self.last_output = []
-        self.lock = lock
-        self.methods = OrderedDict()
-        self.module_class = None
-        self.module_inst = ''.join(module.split(' ')[1:])
-        self.module_name = module.split(' ')[0]
-        #
-        self.load_methods(module, user_modules)
-
-    @staticmethod
-    def load_from_file(filepath):
-        """
-        Return user-written class object from given path.
-        """
-        class_inst = None
-        expected_class = 'Py3status'
-        module_name, file_ext = os.path.splitext(os.path.split(filepath)[-1])
-        if file_ext.lower() == '.py':
-            py_mod = imp.load_source(module_name, filepath)
-            if hasattr(py_mod, expected_class):
-                class_inst = py_mod.Py3status()
-        return class_inst
-
-    @staticmethod
-    def load_from_namespace(module_name):
-        """
-        Load a py3status bundled module.
-        """
-        class_inst = None
-        name = 'py3status.modules.{}'.format(module_name)
-        py_mod = __import__(name)
-        components = name.split('.')
-        for comp in components[1:]:
-            py_mod = getattr(py_mod, comp)
-        class_inst = py_mod.Py3status()
-        return class_inst
-
-    def clear_cache(self):
-        """
-        Reset the cache for all methods of this module.
-        """
-        for meth in self.methods:
-            self.methods[meth]['cached_until'] = time()
-            if self.config['debug']:
-                syslog(LOG_INFO, 'clearing cache for method {}'.format(meth))
-
-    def load_methods(self, module, user_modules):
-        """
-        Read the given user-written py3status class file and store its methods.
-        Those methods will be executed, so we will deliberately ignore:
-            - private methods starting with _
-            - decorated methods such as @property or @staticmethod
-            - 'on_click' methods as they'll be called upon a click_event
-            - 'kill' methods as they'll be called upon this thread's exit
-        """
-        # user provided modules take precedence over py3status provided modules
-        if self.module_name in user_modules:
-            include_path, f_name = user_modules[self.module_name]
-            syslog(LOG_INFO,
-                   'loading module "{}" from {}{}'.format(module, include_path,
-                                                          f_name))
-            class_inst = self.load_from_file(include_path + f_name)
-        # load from py3status provided modules
-        else:
-            syslog(LOG_INFO,
-                   'loading module "{}" from py3status.modules.{}'.format(
-                       module, self.module_name))
-            class_inst = self.load_from_namespace(self.module_name)
-
-        if class_inst:
-            self.module_class = class_inst
-
-            # apply module configuration from i3status config
-            mod_config = self.i3status_thread.config.get(module, {})
-            for config, value in mod_config.items():
-                setattr(self.module_class, config, value)
-
-            # get the available methods for execution
-            for method in sorted(dir(class_inst)):
-                if method.startswith('_'):
-                    continue
-                else:
-                    m_type = type(getattr(class_inst, method))
-                    if 'method' in str(m_type):
-                        if method == 'on_click':
-                            self.click_events = True
-                        elif method == 'kill':
-                            self.has_kill = True
-                        else:
-                            # the method_obj stores infos about each method
-                            # of this module.
-                            method_obj = {
-                                'cached_until': time(),
-                                'instance': None,
-                                'last_output': {
-                                    'name': method,
-                                    'full_text': ''
-                                },
-                                'method': method,
-                                'name': None
-                            }
-                            self.methods[method] = method_obj
-
-        # done, syslog some debug info
-        if self.config['debug']:
-            syslog(LOG_INFO,
-                   'module "{}" click_events={} has_kill={} methods={}'.format(
-                       module, self.click_events, self.has_kill,
-                       self.methods.keys()))
-
-    def click_event(self, event):
-        """
-        Execute the 'on_click' method of this module with the given event.
-        """
-        try:
-            click_method = getattr(self.module_class, 'on_click')
-            click_method(self.i3status_thread.json_list,
-                         self.i3status_thread.config['general'], event)
-        except Exception:
-            err = sys.exc_info()[1]
-            msg = 'on_click failed with ({}) for event ({})'.format(err, event)
-            syslog(LOG_WARNING, msg)
-
-    @profile
-    def run(self):
-        """
-        On a timely fashion, execute every method found for this module.
-        We will respect and set a cache timeout for each method if the user
-        didn't already do so.
-        We will execute the 'kill' method of the module when we terminate.
-        """
-        while self.lock.is_set():
-            # execute each method of this module
-            for meth, obj in self.methods.items():
-                my_method = self.methods[meth]
-
-                # always check the lock
-                if not self.lock.is_set():
-                    break
-
-                # respect the cache set for this method
-                if time() < obj['cached_until']:
-                    continue
-
-                try:
-                    # execute method and get its output
-                    method = getattr(self.module_class, meth)
-                    response = method(self.i3status_thread.json_list,
-                                      self.i3status_thread.config['general'])
-
-                    if isinstance(response, dict):
-                        # this is a shiny new module giving a dict response
-                        result = response
-                    elif isinstance(response, tuple):
-                        # this is an old school module reporting its position
-                        position, result = response
-                        if not isinstance(result, dict):
-                            raise TypeError('response should be a dict')
-                    else:
-                        raise TypeError('response should be a dict')
-
-                    # validate the response
-                    if 'full_text' not in result:
-                        raise KeyError('missing "full_text" key in response')
-                    else:
-                        result['instance'] = self.module_inst
-                        result['name'] = self.module_name
-
-                    # initialize method object
-                    if my_method['name'] is None:
-                        my_method['name'] = result['name']
-                        if 'instance' in result:
-                            my_method['instance'] = result['instance']
-                        else:
-                            my_method['instance'] = result['name']
-
-                    # update method object cache
-                    if 'cached_until' in result:
-                        cached_until = result['cached_until']
-                    else:
-                        cached_until = time() + self.config['cache_timeout']
-                    my_method['cached_until'] = cached_until
-
-                    # update method object output
-                    my_method['last_output'] = result
-
-                    # debug info
-                    if self.config['debug']:
-                        syslog(LOG_INFO,
-                               'method {} returned {} '.format(meth, result))
-                except Exception:
-                    err = sys.exc_info()[1]
-                    syslog(LOG_WARNING,
-                           'user method {} failed ({})'.format(meth, err))
-
-            # don't be hasty mate, let's take it easy for now
-            sleep(self.config['interval'])
-
-        # check and execute the 'kill' method if present
-        if self.has_kill:
-            try:
-                kill_method = getattr(self.module_class, 'kill')
-                kill_method(self.i3status_thread.json_list,
-                            self.i3status_thread.config['general'])
-            except Exception:
-                # this would be stupid to die on exit
-                pass
 
 
 class Py3statusWrapper():
